requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
tensorflow
pillow